% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colMatcher.R
\name{colMatcher}
\alias{colMatcher}
\title{colMatcher}
\usage{
colMatcher(x, return_index = FALSE)
}
\arguments{
\item{x}{The matrix or data.frame}

\item{return_index}{logical.  If TRUE, returns which two columns are
identical.}
}
\value{
If return_index = FALSE, returns TRUE if any columns matched.  If
return_index = TRUE, returns matrix indicating which columns matched.  Always
returns FALSE if no columns matched.
}
\description{
Determines if any two columns of a data.frame or matrix are identical.
}

