% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfSplit.R
\name{dfSplit}
\alias{dfSplit}
\title{Split a data frame into multiple parts.}
\usage{
dfSplit(dat, method = c("kfold", "traintest"), K = NULL,
  train_frac = NULL, train_n = NULL)
}
\arguments{
\item{dat}{Either a data frame or matrix to be split.}

\item{method}{Either \code{"kfold"} or \code{"traintest"}, for splitting
based on K-fold cross validation or a train/test split.}

\item{K}{The number of folds, if \code{"kfold"} is selected.}

\item{train_frac}{The fraction of data to be put into the training set.}

\item{train_n}{The number of observations to be put into the trianing set,
if \code{train_frac} is not selected.}
}
\value{
A list of data frames sliced from the original data.
}
\description{
Given a data frame, split it as either (1) K equal slices for use in K-fold
cross validation or (2) into a train-test split.  If split into a training
and testing set, you may select the training size as either a fraction of the
original data, or by the number of observations.
}

