% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullRanker.R
\name{fullRanker}
\alias{fullRanker}
\title{Convert a matrix to full rank.}
\usage{
fullRanker(x)
}
\arguments{
\item{x}{A matrix or data frame to be converted into a full rank matrix.}
}
\value{
A matrix of full rank that is the original matrix with select columns
removed.
}
\description{
Convert a rank-deficient matrix into a full rank one.
}

