% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdesc.R
\name{mdesc}
\alias{mdesc}
\title{Missing Value Description Table}
\usage{
mdesc(df)
}
\arguments{
\item{df}{A dataframe or vector to analyze.}
}
\value{
An object of type dataframe describing where missing values occur and
at what rate.
}
\description{
Creates a table that describes where missing values occur
in a given dataframe.
}

