% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabstat.R
\name{tabstat}
\alias{tabstat}
\title{Tabstat}
\usage{
tabstat(var, data, by = NULL, fns = list(mean, sd, min, median, max,
  length), na.rm = TRUE)
}
\arguments{
\item{var}{The variable to be summarised.  Must be entered as character.}

\item{data}{The data frame to be summarised.}

\item{by}{A string of character variables to group summaries into.}

\item{fns}{A list of summary functions.  Use `length` for rowcounts by
category. Default already contains many useful statistics.}

\item{na.rm}{Logical, whether or not to remove NAs.  Defaults to true.}
}
\value{
A dataframe of summary statistics.
}
\description{
Creates a 'tabstat' summary statistic table, similar to the ones produced
in Stata.
}

