% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.R
\name{unpack}
\alias{unpack}
\title{Unpacking a list}
\usage{
unpack(list)
}
\arguments{
\item{list}{A list to be unpacked.}
}
\value{
No returns (null).
}
\description{
Unpacks a list into the parent environment.
}

