% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isid.R
\name{isid}
\alias{isid}
\title{Check to see if a column/vector can be used as an id variable.}
\usage{
isid(x)
}
\arguments{
\item{x}{Vector or dataframe.}
}
\value{
logical indicating whether or not `x` has distinct elements.  If a
dataframe was given, a vector of logicals.
}
\description{
Check to see if a column/vector can be used as an id variable.
}

